+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This file is used to list changes made in each version of the resolver cookbook.
 
 ## Unreleased
 
+- Merge resolv.conf options & remove extra colons
+
 ## 3.0.1 - *2021-02-20*
 
 - Update `Chef` -> `Chef Infra`